SIP: Check for all versions 10.11 or later
This abuses the string operator to suppose that future versions will have a
string value which is "greater than" the test value. This is brittle and 
will break after either an OSX/macOS major version bump (unlikely to be of
concern) or after a minor version bump beyond 10.99 (also unlikely to be of
concern).